+++ script/lint.js
@@ -9,6 +9,8 @@ const klaw = require('klaw');
 const minimist = require('minimist');
 const path = require('path');
 
+const { chunkFilenames } = require('./lib/utils');
+
 const ELECTRON_ROOT = path.normalize(path.dirname(__dirname));
 const SOURCE_ROOT = path.resolve(ELECTRON_ROOT, '..');
 const DEPOT_TOOLS = path.resolve(SOURCE_ROOT, 'third_party', 'depot_tools');
@@ -69,12 +71,11 @@ const LINTERS = [{
   roots: ['shell'],
   test: filename => filename.endsWith('.cc') || (filename.endsWith('.h') && !isObjCHeader(filename)),
   run: (opts, filenames) => {
-    if (opts.fix) {
-      spawnAndCheckExitCode('python3', ['script/run-clang-format.py', '-r', '--fix', ...filenames]);
-    } else {
-      spawnAndCheckExitCode('python3', ['script/run-clang-format.py', '-r', ...filenames]);
+    const clangFormatFlags = opts.fix ? ['--fix'] : [];
+    for (const chunk of chunkFilenames(filenames)) {
+      spawnAndCheckExitCode('python3', ['script/run-clang-format.py', ...clangFormatFlags, ...chunk]);
+      cpplint(chunk);
     }
-    cpplint(filenames);
   }
 }, {
   key: 'objc',
